Updated imports to reflect new location of utility class; plugins are no longer outputing anything so commiting for Andrew to take a look at
